requirements.txt
Flask==2.1.2
pandas==1.3.3
scikit-surprise==1.1.1
Flask-CORS==3.0.10
